requirements.txt
openai==0.27.7
python-dotenv==1.0.0
tiktoken==0.4.0